# docs/api/openapi.v1.yaml
openapi: 3.0.3
info:
  title: Alfresco Agent Gateway API
  version: 1.0.0
  description: |
    Hardened API layer in front of alfresco-go-cli for Microsoft 365 Agents SDK integration.
    Enforces strict schema validation, deterministic resolve checks, approval gating for apply,
    and auditable idempotent execution.
servers:
  - url: https://api.example.com
security:
  - bearerAuth: []
...
paths:
  /v1/resolve:
    post:
      tags: [Resolve]
      summary: Resolve target location in Alfresco with deterministic checks
      operationId: resolveTarget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveRequest'
      responses:
        '200':
          description: Resolve response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResolveResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/plan:
    post:
      tags: [Plan]
      summary: Build dry-run plan from selected resolve candidate
      operationId: buildPlan
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanRequest'
      responses:
        '200':
          description: Plan response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/apply:
    post:
      tags: [Apply]
      summary: Request apply execution (approval gated)
      operationId: requestApply
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyRequest'
      responses:
        '202':
          description: Approval required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyResponse'
        '200':
          description: Executed immediately (if policy permits)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/approvals/{approval_id}/decision:
    post:
      tags: [Approvals]
      summary: Approve or reject pending apply request
      operationId: decideApproval
      parameters:
        - name: approval_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalDecisionRequest'
      responses:
        '200':
          description: Approval decision accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalDecisionResponse'
        '404':
          description: Approval not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/approvals/{approval_id}:
    get:
      tags: [Approvals]
      summary: Get approval status
      operationId: getApproval
      parameters:
        - name: approval_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Approval status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalStatusResponse'
        '404':
          description: Approval not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/approvals:
    get:
      tags: [Approvals]
      summary: List approvals
      operationId: listApprovals
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: requested_by
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Approval list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalListResponse'
  /v1/approvals/pending/cards:
    get:
      tags: [Approvals]
      summary: List pending approvals as Teams Adaptive Cards payloads
      operationId: listPendingApprovalCards
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Pending approval cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingApprovalCardsResponse'
  /v1/operations/{operation_id}:
    get:
      tags: [Operations]
      summary: Get operation status and result
      operationId: getOperation
      parameters:
        - name: operation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Operation status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatusResponse'
        '404':
          description: Operation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
  schemas:
    ResolveRequest:
      type: object
      additionalProperties: false
      required: [schema_version, operation, target, scope, policy]
      properties:
        schema_version:
          type: string
          enum: ["1.0"]
        operation:
          type: string
          enum: [locate_for_update, locate_for_upload, locate_for_create]
        scope:
          type: object
          additionalProperties: false
          required: [site_id, root_node_id, max_depth]
          properties:
            site_id: { type: string, minLength: 1 }
            root_node_id: { type: string, format: uuid }
            max_depth: { type: integer, minimum: 1, maximum: 10 }
        target:
          type: object
          additionalProperties: false
          required: [kind, name]
          properties:
            kind: { type: string, enum: [file, folder] }
            name: { type: string, minLength: 1 }
            extension: { type: string }
            expected_parent_path: { type: string }
            expected_site_id: { type: string }
            ancestor_names:
              type: array
              items: { type: string }
            query_text: { type: string }
            modified_within_days: { type: integer, minimum: 0 }
        policy:
          type: object
          additionalProperties: false
          required: [max_candidates, require_unique]
          properties:
            max_candidates: { type: integer, minimum: 1, maximum: 20 }
            require_unique: { type: boolean }
    ResolveResponse:
      type: object
      additionalProperties: false
      required: [schema_version, request_id, status, candidates, confidence, next_action]
      properties:
        schema_version: { type: string, enum: ["1.0"] }
        request_id: { type: string }
        status:
          type: string
          enum: [resolved, ambiguous, not_found, out_of_scope]
        best_candidate_node_id:
          type: string
          format: uuid
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/ResolveCandidate'
        confidence:
          $ref: '#/components/schemas/Confidence'
        next_action: { type: string }
        questions:
          type: array
          items: { type: string }
...
    PlanRequest:
      type: object
      additionalProperties: false
      required: [schema_version, resolve_request_id, action, selection, payload, safety]
      properties:
        schema_version: { type: string, enum: ["1.0"] }
        resolve_request_id: { type: string }
        action:
          type: string
          enum: [upload_new_version, update_metadata, create_child]
        selection:
          type: object
          additionalProperties: false
          properties:
            target_node_id: { type: string, format: uuid }
            target_parent_node_id: { type: string, format: uuid }
        payload:
          type: object
          additionalProperties: false
          properties:
            local_file_path: { type: string }
            new_name: { type: string }
            properties:
              type: object
              additionalProperties:
                type: string
        safety:
          type: object
          additionalProperties: false
          required: [dry_run, require_preconditions]
          properties:
            dry_run: { type: boolean }
            require_preconditions: { type: boolean }
            expected_modified_at: { type: string }
            expected_checksum: { type: string }
...
    ApplyRequest:
      type: object
      additionalProperties: false
      required: [schema_version, plan_id, plan_hash]
      properties:
        schema_version: { type: string, enum: ["1.0"] }
        plan_id: { type: string }
        plan_hash: { type: string }
        idempotency_key: { type: string }
        approved: { type: boolean }
        runtime:
          type: object
          additionalProperties: false
          properties:
            http_timeout: { type: string }
            http_retries: { type: integer, minimum: 0, maximum: 10 }
            http_retry_wait: { type: string }
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, trace_id]
      properties:
        code: { type: string }
        message: { type: string }
        trace_id: { type: string }
        details:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              field: { type: string }
              message: { type: string }